refactor(estree/tokens): break up into multiple files (#19865)
Pure refactor. `serialize.rs` had got too big. Split it up into different files. Does not alter any of the code, just moves it into separate files.
