+++ crates/oxc_estree_tokens/src/lib.rs
@@ -1,87 +1,28 @@
-use oxc_ast::ast::*;
-use oxc_ast_visit::utf8_to_utf16::Utf8ToUtf16;
-use oxc_estree::{CompactFormatter, Config, ESTreeSerializer, PrettyFormatter};
-use oxc_parser::Token;
+//! Functions to prepare tokens for sending to JS side.
+//!
+//! Provides 2 different implementations:
+//!
+//! 1. Serialize tokens to JSON (`json.rs`).
+//! 2. Update `Kind` of tokens in place so they can be deserialized on JS side (`raw_transfer.rs`).
+//!
+//! Both implementations share the same logic by utilizing the `Context` trait and AST visitor `Visitor`.
+//! `Context` trait is implemented by `JsonContext` and `UpdateContext`.
+//! The implementations only differ in how they process tokens sent to them by the visitor.
+//!
+//! Both implementations also convert UTF-8 spans to UTF-16.
 
+mod context;
+mod json;
 mod jsx_state;
 mod options;
-mod serialize;
+mod raw_transfer;
 mod token_type;
 mod u32_string;
+mod visitor;
 
+pub use json::{to_estree_tokens_json, to_estree_tokens_pretty_json};
 pub use jsx_state::{JSXState, JSXStateJS, JSXStateTS};
 pub use options::{
     ESTreeTokenConfig, ESTreeTokenOptions, ESTreeTokenOptionsJS, ESTreeTokenOptionsTS,
 };
-pub use serialize::update_tokens;
-use serialize::{estimate_json_len, serialize_tokens};
-
-/// Serializer config for tokens.
-/// We never include ranges, so use this custom config which returns `false` for `ranges()`.
-/// This allows compiler to remove the branch which checks whether to print ranges in `serialize_span`.
-struct TokenSerializerConfig;
-
-impl Config for TokenSerializerConfig {
-    const INCLUDE_TS_FIELDS: bool = false;
-    const FIXES: bool = false;
-
-    #[expect(clippy::inline_always)] // It's a no-op
-    #[inline(always)]
-    fn new(_ranges: bool) -> Self {
-        Self
-    }
-
-    // Never include ranges, so always return `false`.
-    // `#[inline(always)]` to ensure compiler removes dead code resulting from the static value.
-    #[expect(clippy::inline_always)]
-    #[inline(always)]
-    fn ranges(&self) -> bool {
-        false
-    }
-}
-
-/// Serializer for tokens to compact JSON.
-type CompactTokenSerializer = ESTreeSerializer<TokenSerializerConfig, CompactFormatter>;
-
-/// Serializer for tokens to pretty-printed JSON.
-type PrettyTokenSerializer = ESTreeSerializer<TokenSerializerConfig, PrettyFormatter>;
-
-/// Serialize tokens to JSON.
-///
-/// `program` must have unconverted UTF-8 byte offset spans (as returned by the parser).
-/// Token span conversion to UTF-16 is handled internally.
-///
-/// `source_text` must be the original source text, prior to BOM removal.
-/// i.e. if the file has a BOM, it must be present at the start of `source_text`.
-pub fn to_estree_tokens_json<O: ESTreeTokenConfig>(
-    tokens: &[Token],
-    program: &Program<'_>,
-    source_text: &str,
-    span_converter: &Utf8ToUtf16,
-    options: O,
-) -> String {
-    let capacity = estimate_json_len(tokens.len(), source_text.len(), true);
-    let mut serializer = CompactTokenSerializer::with_capacity(capacity, false);
-    serialize_tokens(&mut serializer, tokens, program, source_text, span_converter, options);
-    serializer.into_string()
-}
-
-/// Serialize tokens to pretty-printed JSON.
-///
-/// `program` must have unconverted UTF-8 byte offset spans (as returned by the parser).
-/// Token span conversion to UTF-16 is handled internally.
-///
-/// `source_text` must be the original source text, prior to BOM removal.
-/// i.e. if the file has a BOM, it must be present at the start of `source_text`.
-pub fn to_estree_tokens_pretty_json<O: ESTreeTokenConfig>(
-    tokens: &[Token],
-    program: &Program<'_>,
-    source_text: &str,
-    span_converter: &Utf8ToUtf16,
-    options: O,
-) -> String {
-    let capacity = estimate_json_len(tokens.len(), source_text.len(), false);
-    let mut serializer = PrettyTokenSerializer::with_capacity(capacity, false);
-    serialize_tokens(&mut serializer, tokens, program, source_text, span_converter, options);
-    serializer.into_string()
-}
+pub use raw_transfer::update_tokens;
